OpenDuet: Re-use DEBUG EfiLdr in NOOPT build
NOOPT EfiLdr is too large to fit in 0x10000-x020000 allotted space, but
we would like to be able to debug everything else in NOOPT. By re-using
DEBUG build EfiLdr in NOOPT, we can.

In fairness, this is a horrible hack - in particular it requires DEBUG
OpenDuet to be built before NOOPT - but it works with the current build
system, and has the arguable advantage of not affecting non-NOOPT builds
at all.