OpenDuet: Fix PeLoader image relocation
There is a gap between the DxeCore and DxeIpl images after this,
but it does not matter because the memory reserved to protect
them both does not add the sizes, but works from start of DxeCore
to end of DxeIpl (based on start and size of DxeIpl, assuming only
that DxeIpl is after DxeCore).

The reason for doing this is to allow the LoadedImageProtocol
for DxeCore to point to the start of the loaded PE image normally,
so that debugging works.

Also includes unecessary (because never used) fix to size in
LoadedImageProtocol which is calculated before EfiLdr handover.